Enable qhs by default if host supports qhs and unicast
Analysis:
When BT is turned on for the first time after flash, persist.vendor.btstack.qhs_enable
is empty before acmService sets it. So, in controller.cc, we will send cmd to disalbe qhs.
The absence of qhs leads to BTFW reported that BN_VARIATION and FT_CHANGE are not supported.
Consequently, this prevents the selection of LE aptX Adaptive.
Fix:
Enable qhs for ISO by default if host supports qhs and LEA unicast.

Change-Id: I37766a25c7325bdb23af77b9898edebbd0fd324b
CRs-Fixed: 3764451